pyproject.toml
[project]
name = "imagedb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer[all]",
    "requests",
    "pillow",
    "pyyaml",
    "lancedb",
]

[project.scripts]
imagedb = "imagedb.main:app"
